# frontend/api-contracts/openapi/client-service.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "OpenAPI definition",
                 "version":  "v0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8200/client-service",
                        "description":  "Generated server url"
                    }
                ],
    "paths":  {
...
                  "/api/verificationTokens/{verificationTokenId}":  {
                                                                        "get":  {
                                                                                    "tags":  [
                                                                                                 "verification-token-resource"
                                                                                             ],
                                                                                    "operationId":  "findById",
                                                                                    "parameters":  [
                                                                                                       {
                                                                                                           "name":  "verificationTokenId",
                                                                                                           "in":  "path",
                                                                                                           "required":  true,
                                                                                                           "schema":  {
                                                                                                                          "type":  "string"
                                                                                                                      }
                                                                                                       }
                                                                                                   ],
                                                                                    "responses":  {
                                                                                                      "200":  {
                                                                                                                  "description":  "OK",
                                                                                                                  "content":  {
                                                                                                                                  "*/*":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "$ref":  "#/components/schemas/VerificationTokenDto"
                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                              }
                                                                                                              }
                                                                                                  }
                                                                                },
                                                                        "put":  {
                                                                                    "tags":  [
                                                                                                 "verification-token-resource"
                                                                                             ],
                                                                                    "operationId":  "update_1",
                                                                                    "parameters":  [
                                                                                                       {
                                                                                                           "name":  "verificationTokenId",
                                                                                                           "in":  "path",
                                                                                                           "required":  true,
                                                                                                           "schema":  {
                                                                                                                          "type":  "string"
                                                                                                                      }
                                                                                                       }
                                                                                                   ],
                                                                                    "requestBody":  {
                                                                                                        "content":  {
                                                                                                                        "application/json":  {
                                                                                                                                                 "schema":  {
                                                                                                                                                                "$ref":  "#/components/schemas/VerificationTokenDto"
                                                                                                                                                            }
                                                                                                                                             }
                                                                                                                    },
                                                                                                        "required":  true
                                                                                                    },
                                                                                    "responses":  {
                                                                                                      "200":  {
                                                                                                                  "description":  "OK",
                                                                                                                  "content":  {
                                                                                                                                  "*/*":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "$ref":  "#/components/schemas/VerificationTokenDto"
                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                              }
                                                                                                              }
                                                                                                  }
                                                                                },
                                                                        "delete":  {
                                                                                       "tags":  [
                                                                                                    "verification-token-resource"
                                                                                                ],
                                                                                       "operationId":  "deleteById",
                                                                                       "parameters":  [
                                                                                                          {
                                                                                                              "name":  "verificationTokenId",
                                                                                                              "in":  "path",
                                                                                                              "required":  true,
                                                                                                              "schema":  {
                                                                                                                             "type":  "string"
                                                                                                                         }
                                                                                                          }
                                                                                                      ],
                                                                                       "responses":  {
                                                                                                         "200":  {
                                                                                                                     "description":  "OK",
                                                                                                                     "content":  {
                                                                                                                                     "*/*":  {
                                                                                                                                                 "schema":  {
                                                                                                                                                                "type":  "boolean"
                                                                                                                                                            }
                                                                                                                                             }
                                                                                                                                 }
                                                                                                                 }
                                                                                                     }
                                                                                   }
                                                                    },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "VerificationTokenDto":  {
                                                                    "type":  "object",
                                                                    "properties":  {
                                                                                       "verificationTokenId":  {
                                                                                                                   "type":  "integer",
                                                                                                                   "format":  "int32"
                                                                                                               },
                                                                                       "token":  {
                                                                                                     "type":  "string"
                                                                                                 },
                                                                                       "expireDate":  {
                                                                                                          "type":  "string",
                                                                                                          "format":  "date"
                                                                                                      },
                                                                                       "credential":  {
                                                                                                          "$ref":  "#/components/schemas/CredentialDto"
                                                                                                      }
                                                                                   }
                                                                },
...
                                   }
                   }
}